+++ src/window/window.hpp
@@ -18,9 +18,9 @@ public:
     void run();
 
 private:
-    void update(double dt);
+    void update(float dt);
     void render();
-    void update_frame_counter(double dt);
+    void update_frame_counter(float dt);
 
     static void key_callback(GLFWwindow* p_window, int key, int scancode, int action, int mods);
